tabs.misc: moved about and error tabs to misc module
